+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Readded video layers' `mediaX`, `mediaY`, `mediaWidth` and `mediaHeight` properties.
 
 ### Changed
+- All movie and layer arguments are now supplied in an options object.
 - Keyframes are now entered as `new vd.KeyFrame([time1, val1, interp],
   [time2, val2])`
 - Rename `cctx` ('canvas context') to `vctx` ('visual context').